requirements.txt
python-dotenv>=1.0.0
aiohttp>=3.9.0
websocket-client>=1.0
requests>=2.28.0